refactor (models): Relocate DQM1 models to dedicated directory
- Moved all DQM1 specific models from `src/app/models.py` to
  `src/app/models/dqm1/` for better organization
- `src/app/models.py` file deleted because it's empty
