+++ src/app/models/dqm1/item.py
@@ -1,0 +1,17 @@
+from typing import Optional
+
+from sqlmodel import Field, SQLModel
+
+
+class Item(SQLModel, table=True):
+    """
+    Lists all items sold in shops and found in the field
+    """
+
+    id: Optional[int] = Field(default=None, primary_key=True)
+    item_name: str
+    item_category: str
+    item_description: str
+    price: Optional[int] = Field(default=None)
+    sell_price: Optional[int] = Field(default=None)
+    sell_location: str
